Move all RNamName stuff into rnams.h
