+++ src/congpairs.cc
@@ -21,9 +21,9 @@
 #include <string>
 #include <vector>
 
-#include "data.h"
-#include "semigrp.h"
 #include "fropin.h"
+#include "rnams.h"
+#include "semigrp.h"
 
 #include "semigroupsplusplus/tc.h"
 
@@ -69,10 +69,11 @@ static void cong_obj_init_cpp_cong(Obj o) {
   gap_semigroup_t range_obj = cong_obj_get_range_obj(o);
   std::string type = std::string(CSTR_STRING(ElmPRec(o, RNam_fin_cong_type)));
   Congruence* cong;
+  bool        report = semi_obj_get_report(range_obj);
 
   if (cong_obj_get_range_type(o) != UNKNOWN) {
     Semigroup* range = cong_obj_get_range(o);
-    range->enumerate(Semigroup::LIMIT_MAX, rec_get_report(o));
+    range->enumerate(Semigroup::LIMIT_MAX, report);
 
     std::vector<relation_t> extra;
     word_t                  lhs, rhs;
@@ -92,12 +93,12 @@ static void cong_obj_init_cpp_cong(Obj o) {
     }
     if (range->size() > 1000) {
       cong = parallel_todd_coxeter(
-          new Congruence(type, range, extra, true, rec_get_report(o), 1),
-          new Congruence(type, range, extra, false, rec_get_report(o), 2),
-          rec_get_report(o));
+          new Congruence(type, range, extra, true, report, 1),
+          new Congruence(type, range, extra, false, report, 2),
+          report);
     } else {
-      cong = new Congruence(type, range, extra, true, rec_get_report(o), 1);
-      cong->todd_coxeter(rec_get_report(o));
+      cong = new Congruence(type, range, extra, true, report, 1);
+      cong->todd_coxeter(report);
     }
   } else {
     gap_prec_t              data  = fropin(range_obj, INTOBJ_INT(-1), 0, False);
@@ -157,11 +158,11 @@ static void cong_obj_init_cpp_cong(Obj o) {
           new Congruence(
               type, nrgens, std::vector<relation_t>(), extra, prefill, 1),
           new Congruence(type, nrgens, rels, extra, 2),
-          rec_get_report(o));
+          report);
     } else {
       cong = new Congruence(
           type, nrgens, std::vector<relation_t>(), extra, prefill, 1);
-      cong->todd_coxeter(rec_get_report(o));
+      cong->todd_coxeter(report);
     }
   }
   cong->compress();
@@ -186,16 +187,14 @@ Obj CONG_PAIRS_NR_CLASSES(Obj self, Obj o) {
 Obj CONG_PAIRS_IN(Obj self, gap_cong_t o, gap_plist_t pair) {
   initRNams();
   gap_semigroup_t S = cong_obj_get_range_obj(o);
-  size_t lhs_pos =
-      INT_INTOBJ(EN_SEMI_POSITION(0L, S, ELM_LIST(pair, 1)));
-  size_t rhs_pos =
-      INT_INTOBJ(EN_SEMI_POSITION(0L, S, ELM_LIST(pair, 2)));
+  size_t lhs_pos    = INT_INTOBJ(EN_SEMI_POSITION(0L, S, ELM_LIST(pair, 1)));
+  size_t rhs_pos    = INT_INTOBJ(EN_SEMI_POSITION(0L, S, ELM_LIST(pair, 2)));
 
-  //FIXME check if l/rhs_pos are undefined!!
+  // FIXME check if l/rhs_pos are undefined!!
   if (IsbPRec(o, RNam_fin_cong_lookup)) {
-    //TODO use FindPRec and GET_ELM_PREC 
+    // TODO use FindPRec and GET_ELM_PREC
     Obj lookup = ElmPRec(o, RNam_fin_cong_lookup);
-    Obj out = True;
+    Obj out    = True;
     EQ_INTOBJS(out, ELM_PLIST(lookup, lhs_pos), ELM_PLIST(lookup, rhs_pos));
     return out;
   }
@@ -227,8 +226,9 @@ Obj CONG_PAIRS_LOOKUP_PART(Obj self, gap_cong_t o) {
     return True;
   }
 
-  Congruence* cong   = cong_obj_get_cpp(o);
-  bool        report = rec_get_report(o);
+  Congruence*     cong      = cong_obj_get_cpp(o);
+  gap_semigroup_t range_obj = cong_obj_get_range_obj(o);
+  bool            report    = semi_obj_get_report(range_obj);
 
   Obj partition = NEW_PLIST(T_PLIST_HOM, cong->nr_classes());
   SET_LEN_PLIST(partition, cong->nr_classes());
@@ -287,29 +287,30 @@ Obj CONG_PAIRS_CLASS_COSET_ID(Obj self, gap_cong_class_t o) {
   initRNams();
   // TODO(JDM) assert o is correct type of object
 
-  Obj rep      = ElmPRec(o, RNam_rep);
-  Obj cong_obj = ElmPRec(o, RNam_cong);
-  Obj S        = cong_obj_get_range_obj(cong_obj);
+  Obj             rep       = ElmPRec(o, RNam_rep);
+  Obj             cong_obj  = ElmPRec(o, RNam_cong);
+  gap_semigroup_t range_obj = cong_obj_get_range_obj(cong_obj);
+  bool            report    = semi_obj_get_report(range_obj);
 
   if (IsbPRec(o, RNam_fin_cong_lookup)) {
     // TODO Use FindPRec and GET_ELM_PREC
     Obj lookup = ElmPRec(cong_obj, RNam_fin_cong_lookup);
-    return ELM_PLIST(lookup, INT_INTOBJ(EN_SEMI_POSITION(self, S, rep)));
+    return ELM_PLIST(lookup,
+                     INT_INTOBJ(EN_SEMI_POSITION(self, range_obj, rep)));
   } else if (cong_obj_get_range_type(cong_obj) != UNKNOWN) {
     Congruence* cong  = cong_obj_get_cpp(cong_obj);
     Semigroup*  range = cong_obj_get_range(cong_obj);
 
     word_t word;
-    range->factorisation(word,
-                         INT_INTOBJ(EN_SEMI_POSITION(self, S, rep)) - 1,
-                         rec_get_report(o));
+    range->factorisation(
+        word, INT_INTOBJ(EN_SEMI_POSITION(self, range_obj, rep)) - 1, report);
     return INTOBJ_INT(cong->word_to_coset(word));
   } else {
-    gap_prec_t data = fropin(S, INTOBJ_INT(-1), 0, False);
+    gap_prec_t  data = fropin(range_obj, INTOBJ_INT(-1), 0, False);
     Congruence* cong = cong_obj_get_cpp(cong_obj);
 
     Obj word = ELM_PLIST(ElmPRec(data, RNam_words),
-                         INT_INTOBJ(EN_SEMI_POSITION(self, S, rep)));
+                         INT_INTOBJ(EN_SEMI_POSITION(self, range_obj, rep)));
 
     return INTOBJ_INT(cong->word_to_coset(plist_to_word_t(word)));
   }
