--- src/ideals.h
@@ -1,28 +1,0 @@
-//
-// Semigroups package for GAP
-// Copyright (C) 2016 James D. Mitchell and Wilf A. Wilson
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
-
-#ifndef SEMIGROUPS_SRC_IDEALS_H_
-#define SEMIGROUPS_SRC_IDEALS_H_
-
-#include "src/compiled.h"
-
-// GAP level functions
-
-Obj IDEAL_SIZE(Obj self, Obj ideal);
-
-#endif // SEMIGROUPS_SRC_IDEALS_H_
